refactor statusbarselectors
