+++ src/components/DataView/helpers/index.js
@@ -1,15 +1,15 @@
-import { LEVEL_COLORS } from "../theme/theme";
-import { TAGS_LEVEL } from "../consts/consts";
+import { LEVEL_COLORS as lColors } from "../theme/theme";
+import { TAGS_LEVEL as tLevel } from "../consts/consts";
 import * as moment from "moment";
 
 export function getRowColor(tags) {
     if (tags?.["level"]) {
-        const level = Object.keys(TAGS_LEVEL).find((level) =>
-            TAGS_LEVEL[level].includes(tags.level.toLowerCase())
+        const level = Object.keys(tLevel).find((level) =>
+            tLevel[level].includes(tags.level.toLowerCase())
         );
-        return LEVEL_COLORS[level];
+        return lColors[level];
     } else {
-        return LEVEL_COLORS["unknown"];
+        return lColors["unknown"];
     }
 }
 
@@ -18,7 +18,7 @@ export function toggleActiveStyles(idx) {
         ? "value-tags-container labelsActive"
         : "value-tags-container labelsInactive";
 }
-
+ 
 export function formatDate(timestamp) {
     return moment(parseInt(timestamp)).format("YYYY-MM-DD HH:mm:ss.SSS UTC");
 }
