+++ src/components/StatusBar/components/copylinkbutton/helpers/index.js
@@ -1,0 +1,8 @@
+import localUrl from "../../../../../services/localUrl";
+
+export function storedUrl() {
+    return localUrl().add({
+        data: window.location.href,
+        description: "From Shared URL",
+    });
+}
