+++ src/scss/base/body.scss
@@ -5,3 +5,6 @@ body {
     padding: 0 auto;
     overflow: hidden;
 }
+html {
+    background:$grey-dark; 
+  }
