+++ src/store/createInitialState.js
@@ -26,7 +26,7 @@ const initialState = () => {
         label:externalState.label || 'Last 5 minutes',
         messages: [],
         limitLoad: false,
-        limit: externalState.limit || 1000,
+        limit: externalState.limit || 100,
         step: externalState.step || 100,
         rangeOpen: false,
         labelsBrowserOpen: true,
